Extract out non firearms mixin
