--- exporter/applications/views/goods/add_good_software/views/mixins.py
@@ -1,9 +1,0 @@
-from django.http import Http404
-from django.conf import settings
-
-
-class NonFirearmsFlagMixin:
-    def dispatch(self, request, **kwargs):
-        if not settings.FEATURE_FLAG_NON_FIREARMS_ENABLED:
-            raise Http404
-        return super().dispatch(request, **kwargs)
